to say I am frustrated due to Handle::from_memory() is a understatement
